app/src/main/java/com/example/appevents/data/EventsApi.kt
package com.example.appevents.data

import com.example.appevents.model.CheckIn
import com.example.appevents.model.Event
import retrofit2.Response
import retrofit2.http.Body
import retrofit2.http.GET
import retrofit2.http.POST
import retrofit2.http.Path

interface EventsApi {

    @GET("events/")
    suspend fun getEvents(): Response<ArrayList<Event>>

    @GET("events/{id}")
    suspend fun getEventDetail(@Path("id") id: String): Event

    @POST("checkin")
    suspend fun checkInEvent(@Body checkIn: CheckIn): Response<Any>

}